Fixed large int declerations.
Accidentally declared near everything as "int" and I believe this lead
to insane memory ussage so I cut down almost all of it.

+++ WAndersonLab6/Core/Inc/game_state.h
@@ -12,9 +12,9 @@
 #include "LCD_Driver.h"///////////////// MAYBE NEED TO DELETE
 
 
-int placeCoin(int board[6][7], int col, int player);
+int placeCoin(uint8_t board[6][7], uint8_t col, uint8_t player);
 
-int checkWinOrTie(int board[6][7], int row, int col, int player);
+int checkWinOrTie(uint8_t board[6][7], uint8_t row, uint8_t col, uint8_t player);
 
 
 #endif /* INC_GAME_STATE_H_ */
